document-verifier/backend/requirements.txt
Flask
flask-cors
mysql-connector-python
pymongo
pillow
pytesseract